scheduler: split service/batch alloc logic into its own package.
